Move ModeHandler._renderStatusBar() logic into the more generally-available StatusBar.Clear()
Actions such as `dd` which would totally clear the status bar before now correctly reset it to show the mode, macro being recorded, etc.
Fixes #3838
